# Add in-process concurrency support to the PS worker