--- a/src/Logging/RpcLogger.cs
+++ b/src/Logging/RpcLogger.cs
@@ -12,7 +12,7 @@ using LogLevel = Microsoft.Azure.WebJobs.Script.Grpc.Messages.RpcLog.Types.Level
 
 namespace Microsoft.Azure.Functions.PowerShellWorker.Utility
 {
-    internal class RpcLogger : ILogger, IDisposable
+    internal class RpcLogger : ILogger
     {
         private const string SystemLogPrefix = "LanguageWorkerConsoleLog";
         private readonly MessagingStream _msgStream;
@@ -20,20 +20,19 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.Utility
         private string _invocationId;
         private string _requestId;
 
-        public RpcLogger(MessagingStream msgStream)
+        internal RpcLogger(MessagingStream msgStream)
         {
             _msgStream = msgStream;
             _systemLogMsg = new StringBuilder();
         }
 
-        public IDisposable BeginScope(string requestId, string invocationId)
+        public void SetContext(string requestId, string invocationId)
         {
             _requestId = requestId;
             _invocationId = invocationId;
-            return this;
         }
 
-        public void Dispose()
+        public void ResetContext()
         {
             _requestId = null;
             _invocationId = null;