--- a/src/Messaging/MessagingStream.cs
+++ b/src/Messaging/MessagingStream.cs
@@ -4,6 +4,7 @@
 //
 
 using System;
+using System.Collections.Concurrent;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -14,31 +15,42 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.Messaging
 {
     internal class MessagingStream
     {
-        private SemaphoreSlim _writeSemaphore = new SemaphoreSlim(1, 1);
-        private AsyncDuplexStreamingCall<StreamingMessage, StreamingMessage> _call;
+        private readonly AsyncDuplexStreamingCall<StreamingMessage, StreamingMessage> _call;
+        private readonly BlockingCollection<StreamingMessage> _msgQueue;
 
-        public MessagingStream(string host, int port)
+        internal MessagingStream(string host, int port)
         {
             Channel channel = new Channel(host, port, ChannelCredentials.Insecure);
             _call = new FunctionRpc.FunctionRpcClient(channel).EventStream();
+
+            _msgQueue = new BlockingCollection<StreamingMessage>();
+            Task.Run(WriteImplAsync);
         }
 
-        public StreamingMessage GetCurrentMessage() => _call.ResponseStream.Current;
+        /// <summary>
+        /// Get the current message.
+        /// </summary>
+        internal StreamingMessage GetCurrentMessage() => _call.ResponseStream.Current;
 
-        public async Task<bool> MoveNext() => await _call.ResponseStream.MoveNext(CancellationToken.None);
+        /// <summary>
+        /// Move to the next message.
+        /// </summary>
+        internal async Task<bool> MoveNext() => await _call.ResponseStream.MoveNext(CancellationToken.None);
 
-        public async Task WriteAsync(StreamingMessage message)
+        /// <summary>
+        /// Write the outgoing message to the buffer.
+        /// </summary>
+        internal void Write(StreamingMessage message) => _msgQueue.Add(message);
+
+        /// <summary>
+        /// Take a message from the buffer and write to the gRPC channel.
+        /// </summary>
+        private async Task WriteImplAsync()
         {
-            try
+            while (true)
             {
-                // Wait for the handle to be released because we can't have
-                // more than one message being sent at the same time
-                await _writeSemaphore.WaitAsync();
-                await _call.RequestStream.WriteAsync(message);
-            }
-            finally
-            {
-                _writeSemaphore.Release();
+                StreamingMessage msg = _msgQueue.Take();
+                await _call.RequestStream.WriteAsync(msg);
             }
         }
     }