--- a/src/PowerShell/PowerShellManagerPool.cs
+++ b/src/PowerShell/PowerShellManagerPool.cs
@@ -4,7 +4,11 @@
 //
 
 using System;
+using System.Collections.Concurrent;
+using System.Threading;
+using Microsoft.Azure.Functions.PowerShellWorker.Messaging;
 using Microsoft.Azure.Functions.PowerShellWorker.Utility;
+using Microsoft.Azure.WebJobs.Script.Grpc.Messages;
 
 namespace Microsoft.Azure.Functions.PowerShellWorker.PowerShell
 {
@@ -15,16 +19,18 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.PowerShell
     /// </summary>
     internal class PowerShellManagerPool
     {
-        private readonly ILogger _logger;
-        // Today we don't really support the in-proc concurrency. We just hold an instance of PowerShellManager in this field.
-        private PowerShellManager _psManager;
+        private const int UpperBound = 25;
+        private readonly MessagingStream _msgStream;
+        private readonly BlockingCollection<PowerShellManager> _pool;
+        private int _poolSize;
 
         /// <summary>
         /// Constructor of the pool.
         /// </summary>
-        internal PowerShellManagerPool(ILogger logger)
+        internal PowerShellManagerPool(MessagingStream msgStream)
         {
-            _logger = logger;
+            _msgStream = msgStream;
+            _pool = new BlockingCollection<PowerShellManager>(UpperBound);
         }
 
         /// <summary>
@@ -32,20 +38,57 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.PowerShell
         /// When it's time to really implement this pool, we probably should instantiate PowerShellManager instances in a lazy way.
         /// Maybe start from size 1 and increase the number of workers as needed.
         /// </summary>
-        internal void Initialize()
+        internal void Initialize(string requestId)
         {
-            _psManager = new PowerShellManager(_logger);
+            var logger = new RpcLogger(_msgStream);
+
+            try
+            {
+                logger.SetContext(requestId, invocationId: null);
+                _pool.Add(new PowerShellManager(logger));
+                _poolSize = _pool.Count;
+            }
+            finally
+            {
+                logger.ResetContext();
+            }
         }
 
         /// <summary>
         /// Checkout an idle PowerShellManager instance.
         /// When it's time to really implement this pool, this method is supposed to block when there is no idle instance available.
         /// </summary>
-        internal PowerShellManager CheckoutIdleWorker(AzFunctionInfo functionInfo)
+        internal PowerShellManager CheckoutIdleWorker(StreamingMessage request, AzFunctionInfo functionInfo)
         {
+            PowerShellManager psManager = null;
+            string requestId = request.RequestId;
+            string invocationId = request.InvocationRequest?.InvocationId;
+
+            // If the pool has an idle one, just use it.
+            if (!_pool.TryTake(out psManager))
+            {
+                // The pool doesn't have an idle one.
+                if (_poolSize < UpperBound &&
+                    Interlocked.Increment(ref _poolSize) <= UpperBound)
+                {
+                    // If the pool hasn't reached its bounded capacity yet, then
+                    // we create a new item and return it.
+                    var logger = new RpcLogger(_msgStream);
+                    logger.SetContext(requestId, invocationId);
+                    psManager = new PowerShellManager(logger);
+                }
+                else
+                {
+                    // If the pool has reached its bounded capacity, then the thread
+                    // should be blocked until an idle one becomes available.
+                    psManager = _pool.Take();
+                }
+            }
+
             // Register the function with the Runspace before returning the idle PowerShellManager.
-            FunctionMetadata.RegisterFunctionMetadata(_psManager.InstanceId, functionInfo);
-            return _psManager;
+            FunctionMetadata.RegisterFunctionMetadata(psManager.InstanceId, functionInfo);
+            psManager.Logger.SetContext(requestId, invocationId);
+            return psManager;
         }
 
         /// <summary>
@@ -57,6 +100,8 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.PowerShell
             {
                 // Unregister the Runspace before reclaiming the used PowerShellManager.
                 FunctionMetadata.UnregisterFunctionMetadata(psManager.InstanceId);
+                psManager.Logger.ResetContext();
+                _pool.Add(psManager);
             }
         }
     }