… src/Utility/ExecutionTimer.cs
… src/Worker.cs
… test/Microsoft.Azure.Functions.PowerShellWorker.Test.csproj
… test/Unit/Logging/ConsoleLogger.cs
… test/Unit/Requests/ProcessWorkerInitRequestTests.cs
… test/Unit/xunit.runner.json