@daxian-dbw 2019-01-07
Thanks for the review, the in-proc concurrency support is now merged.